fix: better schema imports
