+++ package.json
@@ -50,7 +50,6 @@
     "mocha-junit-reporter": "2.0.2",
     "mocha-multi-reporters": "1.5.1",
     "nock": "13.2.1",
-    "proxyquire": "2.1.3",
     "remark-stringify": "10.0.2",
     "semantic-release": "18.0.1",
     "sinon": "12.0.1",
