+++ src/defaults/html.pipe.js
@@ -20,14 +20,14 @@ import rewriteBlobImages from '../html/rewrite-blob-images.js';
 import rewriteIcons from '../html/rewrite-icons.js';
 import meta from '../html/get-metadata.js';
 import html from '../html/make-html.js';
-import type from '../utils/set-content-type.js';
+import type from '../html/set-content-type.js';
 import selectStatus from '../html/set-status.js';
 import smartypants from '../html/smartypants.js';
 import sections from '../html/split-sections.js';
-import { selectstrain, selecttest } from '../utils/conditional-sections.js';
+import { selectstrain, selecttest } from '../html/conditional-sections.js';
 import debug from '../html/output-debug.js';
 import { esi, flag } from '../html/flag-esi.js';
-import { adjustHTML, adjustMDAST } from '../utils/adjust-markup.js';
+import { adjustHTML, adjustMDAST } from '../html/adjust-markup.js';
 import dataSections from '../html/data-sections.js';
 import dataEmbeds from '../html/fetch-data.js';
 import removeHlxProps from '../html/removeHlxProps.js';
@@ -39,8 +39,8 @@ import rewriteLinks from '../html/static-asset-links.js';
 import unwrapSoleImages from '../html/unwrap-sole-images.js';
 import embeds from '../html/find-embeds.js';
 import { cache, uncached } from '../html/shared-cache.js';
-import validate from '../utils/validate.js';
-import { record, report } from '../utils/dump-context.js';
+import validate from '../html/validate.js';
+import { record, report } from '../html/dump-context.js';
 import production from '../utils/is-production.js';
 import lastModified from '../html/set-last-modified.js';
 import key from '../html/set-surrogate-key.js';
