+++ src/defaults/json.pipe.js
@@ -13,12 +13,12 @@ import { Pipeline } from '../../index.js';
 import timing from '../utils/timing.js';
 import { selectStatus } from '../json/set-json-status.js';
 import emit from '../json/emit-json.js';
-import validate from '../utils/validate.js';
-import { record, report } from '../utils/dump-context.js';
+import validate from '../html/validate.js';
+import { record, report } from '../html/dump-context.js';
 import production from '../utils/is-production.js';
 import sections from '../html/split-sections.js';
 import smartypants from '../html/smartypants.js';
-import type from '../utils/set-content-type.js';
+import type from '../html/set-content-type.js';
 import meta from '../html/get-metadata.js';
 import parse from '../html/parse-markdown.js';
 import fetchContent from '../html/fetch-content.js';
