+++ src/html/coerce-secrets.js
@@ -1,0 +1,22 @@
+/*
+ * Copyright 2021 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+import ajv from '../utils/validator.js';
+
+export default function coerce(action) {
+  const defaultsetter = ajv(action.logger, { useDefaults: true, coerceTypes: true });
+  if (!action.secrets) {
+    action.secrets = {};
+  }
+  action.logger.debug('Coercing secrets');
+  defaultsetter.validate('https://ns.adobe.com/helix/pipeline/secrets', action.secrets);
+  return action;
+}
