+++ src/html/dump-context.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Adobe. All rights reserved.
+ * Copyright 2021 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -12,7 +12,7 @@
 import path from 'path';
 import fs from 'fs-extra';
 import { setdefault } from 'ferrum';
-import isProduction from './is-production.js';
+import isProduction from '../utils/is-production.js';
 
 /**
  * Returns {@code true} if context dumps should never be written to disk.
