+++ src/schemas/mdast.schema.cjs
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Adobe. All rights reserved.
+ * Copyright 2021 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -9,12 +9,4 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-import { setdefault } from 'ferrum';
-
-export default function type(mime) {
-  return function setmime(context) {
-    const res = setdefault(context, 'response', {});
-    const head = setdefault(res, 'headers', {});
-    setdefault(head, 'Content-Type', mime);
-  };
-}
+module.exports = require('./mdast.schema.json');
