+++ src/utils/validator.js
@@ -10,10 +10,10 @@
  * governing permissions and limitations under the License.
  */
 /* eslint-disable no-underscore-dangle */
-import { createRequire } from 'module';
 import Ajv from 'ajv';
 import hash from 'object-hash';
 import util from 'util';
+import SCHEMAS from '../schemas/index.cjs';
 
 const _ajv = {};
 
@@ -28,31 +28,7 @@ export default function ajv(logger, options = {}) {
       validateFormats: false,
       ...options,
     });
-    // ajvFormats(validator);
-    // compromise: in order to avoid async code here
-    // (which would complicate pipeline implementation considerably)
-    // we're using static file names and synchronous reads/requires (#134)
-    // using constants in the require functions allows packagers to include the schemas.
-    const require = createRequire(import.meta.url);
-    [
-      /* eslint-disable global-require */
-      require('../schemas/action.schema.json'),
-      require('../schemas/content.schema.json'),
-      require('../schemas/context.schema.json'),
-      require('../schemas/mdast.schema.json'),
-      require('../schemas/meta.schema.json'),
-      require('../schemas/position.schema.json'),
-      require('../schemas/rawrequest.schema.json'),
-      require('../schemas/request.schema.json'),
-      require('../schemas/response.schema.json'),
-      require('../schemas/secrets.schema.json'),
-      require('../schemas/section.schema.json'),
-      require('../schemas/textcoordinates.schema.json'),
-      require('@adobe/helix-shared-config/src/schemas/markupconfig.schema.json'),
-      require('@adobe/helix-shared-config/src/schemas/markup.schema.json'),
-      require('@adobe/helix-shared-config/src/schemas/markupmapping.schema.json'),
-      /* eslint-enable global-require */
-    ].forEach((schemaData) => {
+    SCHEMAS.forEach((schemaData) => {
       validator.addSchema(schemaData);
       logger.debug(`- ${schemaData.$id}`);
     });
