+++ test/fetch-data.test.js
@@ -19,7 +19,7 @@ import nock from 'nock';
 import { JSDOM } from 'jsdom';
 import { pipe } from '../src/defaults/html.pipe.js';
 import { resolver } from './utils.js';
-import coerce from '../src/utils/coerce-secrets.js';
+import coerce from '../src/html/coerce-secrets.js';
 import Downloader from '../src/utils/Downloader.js';
 
 const params = {
