+++ test/html.pipe.content.test.js
@@ -14,7 +14,7 @@ import assert from 'assert';
 import { logging } from '@adobe/helix-testutils';
 import nock from 'nock';
 import { pipe } from '../src/defaults/html.pipe.js';
-import coerce from '../src/utils/coerce-secrets.js';
+import coerce from '../src/html/coerce-secrets.js';
 import Downloader from '../src/utils/Downloader.js';
 
 describe('Testing fetch content', () => {
